Cleanup: remove old interface, move RNTuple tests to new file
